+++ src/Data/Morpheus/Resolve/Resolve.hs
@@ -11,6 +11,7 @@ module Data.Morpheus.Resolve.Resolve
   , streamResolver
   , statefulResolver
   , RootResCon
+  , fullSchema
   ) where
 
 import           Control.Monad.Trans.Except                (ExceptT (..), runExceptT)
@@ -128,4 +129,11 @@ fullSchema _ = querySchema >>= mutationSchema >>= subscriptionSchema
     maybeOperator fields = Just . operatorType fields
     -- operatorType :: [a] -> Text -> (Text, DataType [a])
     operatorType typeData typeName =
-      (typeName, DataType {typeData, typeName, typeFingerprint = SystemFingerprint typeName, typeDescription = ""})
+      ( typeName
+      , DataType
+          { typeData
+          , typeVisibility = True
+          , typeName
+          , typeFingerprint = SystemFingerprint typeName
+          , typeDescription = ""
+          })