+++ src/Data/Morpheus/Schema/Directive.hs
@@ -10,7 +10,7 @@ module Data.Morpheus.Schema.Directive
 import           Data.Morpheus.Kind                                (KIND, OBJECT)
 import           Data.Morpheus.Schema.DirectiveLocation            (DirectiveLocation)
 import           Data.Morpheus.Schema.Internal.RenderIntrospection (InputValue)
-import           Data.Morpheus.Types.GQLType                       (GQLType (__typeName))
+import           Data.Morpheus.Types.GQLType                       (GQLType (__typeName, __typeVisibility))
 import           Data.Text                                         (Text)
 import           GHC.Generics                                      (Generic)
 
@@ -18,6 +18,7 @@ type instance KIND Directive = OBJECT
 
 instance GQLType Directive where
   __typeName = const "__Directive"
+  __typeVisibility = const False
 
 data Directive = Directive
   { name        :: Text