+++ src/Data/Morpheus/Schema/Type.hs
@@ -13,7 +13,7 @@ import           Data.Morpheus.Schema.EnumValue  (EnumValue)
 import qualified Data.Morpheus.Schema.Field      as F (Field (..))
 import qualified Data.Morpheus.Schema.InputValue as I (InputValue (..))
 import           Data.Morpheus.Schema.TypeKind   (TypeKind)
-import           Data.Morpheus.Types.GQLType     (GQLType (__typeName))
+import           Data.Morpheus.Types.GQLType     (GQLType (__typeName, __typeVisibility))
 import           Data.Text                       (Text)
 import           GHC.Generics                    (Generic)
 
@@ -21,6 +21,7 @@ type instance KIND Type = OBJECT
 
 instance GQLType Type where
   __typeName = const "__Type"
+  __typeVisibility = const False
 
 data Type = Type
   { kind          :: TypeKind