+++ src/Data/Morpheus/Types/ID.hs
@@ -1,4 +1,3 @@
-{-# LANGUAGE DeriveAnyClass    #-}
 {-# LANGUAGE DeriveGeneric     #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE TypeFamilies      #-}
@@ -9,7 +8,7 @@ module Data.Morpheus.Types.ID
 
 import           Data.Morpheus.Kind                 (KIND, SCALAR)
 import           Data.Morpheus.Types.GQLScalar      (GQLScalar (..))
-import           Data.Morpheus.Types.GQLType        (GQLType)
+import           Data.Morpheus.Types.GQLType        (GQLType (__typeVisibility))
 import           Data.Morpheus.Types.Internal.Value (ScalarValue (..))
 import           Data.Text                          (Text, pack)
 import           GHC.Generics                       (Generic)
@@ -21,7 +20,10 @@ type instance KIND ID = SCALAR
 -- serialized always as 'String'
 newtype ID = ID
   { unpackID :: Text
-  } deriving (Generic, GQLType)
+  } deriving (Generic)
+
+instance GQLType ID where
+  __typeVisibility _ = False
 
 instance GQLScalar ID where
   parseValue (Int x)    = return (ID $ pack $ show x)